+++ src/nvm.h
@@ -34,9 +34,9 @@
 /**
  * 	Module version
  */
-#define NVM_VER_MAJOR		( 1 )
+#define NVM_VER_MAJOR		( 2 )
 #define NVM_VER_MINOR		( 0 )
-#define NVM_VER_DEVELOP		( 1 )
+#define NVM_VER_DEVELOP		( 0 )
 
 /**
  * 	Status
@@ -54,6 +54,7 @@ typedef enum
 typedef struct nvm_mem_driver_s
 {
 	nvm_status_t (*pf_nvm_init)		(void);
+	nvm_status_t (*pf_nvm_deinit)	(void);
 	nvm_status_t (*pf_nvm_write)	(const uint32_t addr, const uint32_t size, const uint8_t * const p_data);
 	nvm_status_t (*pf_nvm_read)		(const uint32_t addr, const uint32_t size, uint8_t * const p_data);
 	nvm_status_t (*pf_nvm_erase)	(const uint32_t addr, const uint32_t size);
@@ -74,7 +75,8 @@ typedef struct nvm_region_s
 // Functions Prototypes
 ////////////////////////////////////////////////////////////////////////////////
 nvm_status_t 	nvm_init	(void);
-bool			nvm_is_init	(void);
+nvm_status_t    nvm_deinit  (void);
+nvm_status_t    nvm_is_init	(bool * const p_is_init);
 nvm_status_t 	nvm_write	(const nvm_region_name_t region, const uint32_t addr, const uint32_t size, const uint8_t * const p_data);
 nvm_status_t 	nvm_read	(const nvm_region_name_t region, const uint32_t addr, const uint32_t size, uint8_t * const p_data);
 nvm_status_t 	nvm_erase	(const nvm_region_name_t region, const uint32_t addr, const uint32_t size);