+++ CHANGE_LOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project/module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## V2.0.0 - 14.12.2022
+
+### Changed
+ - API change:
+    + Added new API function: *nvm_deinit*
+    + Change function prototype of *nvm_is_init*
+ - Memory driver interface must contain also de-init function
+
+---
 ## V1.0.1 - 25.07.2021
 
 ### Added